refactor(reproducer): split _core.py to stay under 512-line budget
The merged streamplot + axis-scale + imshow replay additions pushed
_core.py to 539 lines. Extract two cohesive concerns into new modules:

- _finalize_axes.py: the per-axes post-replay finalization passes
  (tick/special-plot finalizers, imshow tick-suppression, recorded-limit
  reapply). Ordering is preserved and documented inline.
- _recipe_info.py: get_recipe_info() metadata inspection (distinct from
  reproduction). Re-exported via _reproducer/__init__.py so all existing
  import paths (figrecipe.utils, figrecipe._api._public) keep working.

_core.py: 539 -> 480 lines. No behavior change.
